On the media page, some of the pictures are a different size. Should I stretch the image to fill the container? or just leave it with some space at the bottom? If I stretched them, they would still be to scale when you clicked on them and looked at them within the slideshow.
I know that some of the images are rotated wrong on the media page. I still have to rotate them (which is a pain in the ass because you have to change to exif data stuff for the image, another reason why web development is retarded).

I also know that the videos don't work in the slideshow. I still have to get the javascript to transition between the two without starting over, so ignore that for now. Also, sometimes the videos all show up, sometimes they don't. I don't know why. And sometimes when I open the media page I get the "aw snap something went wrong with displaying the webpage".

I'm going to take the controls off of the videos and just put the "play" symbol over them so people know they're vidoes but can't actually play them until you click on them for the slideshow.

I still don't really know what to do with the home page but I did something.

I copied the team text on the homepage straight off of Titaniums website, I'll reword it later.

Should we have pictures on the home page? I think it's a good idea, but only if we have pictures that you can see when they're small. The team picture looks bad there. I don't know what to put on the home page.

I added an awards page and a history page. On the history page, are those descriptions too long? I copied them off of their respective game manuals, but I can shorten them up if needed. I did them for as far back as I could find png game logos.